Seperated CatModule into the component folder. Applied some Styling with Hamza's code
